jyotish_kb/requirements.txt
llama-index>=0.12.0,<0.13.0
llama-index-llms-ollama>=0.5.0
llama-index-embeddings-ollama>=0.5.0
llama-index-vector-stores-chroma>=0.4.0
chromadb>=0.6.0
pyyaml>=6.0
click>=8.1.0
streamlit>=1.32.0
